rename reduction_ops to sum
