+++ tfjs-core/src/gradients/Atan2_grad.ts
@@ -21,9 +21,9 @@ import {add} from '../ops/add';
 import {assertAndGetBroadcastShape, getReductionAxes} from '../ops/broadcast_util';
 import {div} from '../ops/div';
 import {mul} from '../ops/mul';
-import {sum} from '../ops/reduction_ops';
 import {reshape} from '../ops/reshape';
 import {square} from '../ops/square';
+import {sum} from '../ops/sum';
 import {neg} from '../ops/unary_ops';
 import {Tensor} from '../tensor';
