+++ tfjs-core/src/gradients/FusedBatchNorm_grad.ts
@@ -19,9 +19,9 @@ import {GradConfig, NamedAttrMap} from '../kernel_registry';
 import {add} from '../ops/add';
 import {getReductionAxes} from '../ops/broadcast_util';
 import {mul} from '../ops/mul';
-import {sum} from '../ops/reduction_ops';
 import {reshape} from '../ops/reshape';
 import {sub} from '../ops/sub';
+import {sum} from '../ops/sum';
 import {scalar} from '../ops/tensor_ops';
 import {tile} from '../ops/tile';
 import {rsqrt} from '../ops/unary_ops';
