+++ tfjs-core/src/gradients/Mod_grad.ts
@@ -20,8 +20,8 @@ import {GradConfig} from '../kernel_registry';
 import {assertAndGetBroadcastShape, getReductionAxes} from '../ops/broadcast_util';
 import {div} from '../ops/div';
 import {mul} from '../ops/mul';
-import {sum} from '../ops/reduction_ops';
 import {reshape} from '../ops/reshape';
+import {sum} from '../ops/sum';
 import {floor, neg} from '../ops/unary_ops';
 import {Tensor} from '../tensor';
