+++ tfjs-core/src/gradients/Pow_grad.ts
@@ -21,9 +21,9 @@ import * as broadcast_util from '../ops/broadcast_util';
 import {greater} from '../ops/greater';
 import {mul} from '../ops/mul';
 import {pow} from '../ops/pow';
-import {sum} from '../ops/reduction_ops';
 import {reshape} from '../ops/reshape';
 import {sub} from '../ops/sub';
+import {sum} from '../ops/sum';
 import {scalar, zerosLike} from '../ops/tensor_ops';
 import {log} from '../ops/unary_ops';
 import {where} from '../ops/where';
