+++ tfjs-core/src/gradients/Prelu_grad.ts
@@ -19,8 +19,8 @@ import {GradConfig} from '../kernel_registry';
 import {getReductionAxes} from '../ops/broadcast_util';
 import {greater} from '../ops/greater';
 import {mul} from '../ops/mul';
-import {sum} from '../ops/reduction_ops';
 import {reshape} from '../ops/reshape';
+import {sum} from '../ops/sum';
 import {zerosLike} from '../ops/tensor_ops';
 import {where} from '../ops/where';
 import {Tensor} from '../tensor';
