+++ tfjs-core/src/ops/compute_weighted_loss.ts
@@ -25,7 +25,7 @@ import {mean} from './mean';
 import {mul} from './mul';
 import {notEqual} from './not_equal';
 import {op} from './operation';
-import {sum} from './reduction_ops';
+import {sum} from './sum';
 import {ones, scalar} from './tensor_ops';
 
 /**
