+++ tfjs-core/src/ops/norm.ts
@@ -25,9 +25,9 @@ import {max} from './max';
 import {min} from './min';
 import {op} from './operation';
 import {pow} from './pow';
-import {sum} from './reduction_ops';
 import {reshape} from './reshape';
 import {square} from './square';
+import {sum} from './sum';
 import {scalar} from './tensor_ops';
 import {abs, sqrt} from './unary_ops';
