+++ tfjs-core/src/tests.ts
@@ -129,7 +129,6 @@ import './ops/rand_test';
 import './ops/random_gamma_test';
 import './ops/random_normal_test';
 import './ops/random_uniform_test';
-import './ops/reduction_ops_test';
 import './ops/relu6_test';
 import './ops/relu_test';
 import './ops/resize_bilinear_test';
@@ -154,6 +153,7 @@ import './ops/spectral_ops_test';
 import './ops/stack_test';
 import './ops/strided_slice_test';
 import './ops/sub_test';
+import './ops/sum_test';
 import './ops/tile_test';
 import './ops/topk_test';
 import './ops/transpose_test';
